[backend] refactor: JPA 호환성을 위해 모든 ID 타입을 long으로 수정

+++ backend/src/main/java/com/example/demo/controller/MemberController.java
@@ -44,7 +44,7 @@ public class MemberController {
 
     // 멤버 삭제 (초대 거절)
     @DeleteMapping("/members/{id}")
-    public ResponseEntity<String> deleteMember(@PathVariable(required = true) int id, @RequestParam(defaultValue = "succ") String succMsg) {
+    public ResponseEntity<String> deleteMember(@PathVariable(required = true) long id, @RequestParam(defaultValue = "succ") String succMsg) {
         memberService.deleteMember(id);
 
         return ResponseEntity.ok("invite refused");
