+++ backend/src/main/java/com/example/demo/dto/MemberInviteInTeamRequestDTO.java
@@ -6,9 +6,9 @@ import lombok.Setter;
 @Getter
 @Setter
 public class MemberInviteInTeamRequestDTO {
-    private int id;
-    private int userId;
-    private int teamId;
-    private int status;
-    private int inviterId;    // 초대를 한 사람
+    private long id;
+    private long userId;
+    private long teamId;
+    private long status;
+    private long inviterId;    // 초대를 한 사람
 }
